Made class edition admin-exclusive

+++ server/routes/Class.js
@@ -25,12 +25,18 @@ router.get("/byId/:id", async (req, res) => {
 })
 
 router.delete("/byId/:id/edit", async (req, res) => {
+    // Block this route for non-admin users
+    if (!await Auth.isAdmin(req.query.uuid, req.params.id)) return
+
     const id = req.params.id
     await Class.destroy({where: {id: id}})
     res.json({message: "Class deleted"})
 })
 
 router.put("/byId/:id/edit", async (req, res) => {
+    // Block this route for non-admin users
+    if (!await Auth.isAdmin(req.body.uuid, req.params.id)) return
+
     const classInfo = req.body
     const id = req.params.id
     console.log(id)
@@ -39,6 +45,9 @@ router.put("/byId/:id/edit", async (req, res) => {
 })
 
 router.post("/:uuid/enroll-to/:id", async(req,res) =>{
+    // Block this route for non-admin users
+    if (!await Auth.isAdmin(req.params.uuid, req.params.id)) return
+
     const classId = req.params.id
     const userId = Auth.getUserId(req.params.uuid).id
     const [entry, created] = await IsInClass.findOrCreate({where:{userId: userId, classId: Number(classId)}});
